Self-Supervised-Scene-Flow-Estimation/requirements.txt
absl-py==0.8.0
apptools==4.5.0
astor==0.8.0
attrs==19.2.0
backcall==0.1.0
bleach==3.1.0
cachetools==4.0.0
certifi==2019.11.28
cffi==1.12.3
chardet==3.0.4
configobj==5.0.6
cvlib==0.2.4
cycler==0.10.0
Cython==0.29.16
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
envisage==4.9.2
filelock==3.0.12
gast==0.3.0
grpcio==1.23.0
h5py==2.10.0
idna==2.9
imageio==2.6.1
imutils==0.5.3
ipdb==0.12.2
ipydatawidgets==4.0.1
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipyvolume==0.5.2
ipywebrtc==0.5.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.1
joblib==0.14.1
json5==0.8.5
jsonschema==3.0.2
jupyter-client==5.3.3
jupyter-core==4.5.0
jupyterlab==1.1.4
jupyterlab-server==1.0.6
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
llvmlite==0.29.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib
mayavi==4.7.1
mistune==0.8.4
nbconvert==5.6.0
nbformat==4.4.0
networkx==2.4
nose==1.3.7
notebook==6.0.1
numba==0.38.1
numpy==1.17.2
open3d==0.8.0.0
open3d-python==0.7.0.0
opencv-python==4.1.1.26
pandas==0.25.2
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
pptk==0.1.0
progressbar==2.5
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.9.1
ptyprocess==0.6.0
pycparser==2.19
pyface==6.1.2
Pygments==2.4.2
pyparsing==2.4.2
pypng==0.0.20
pyquaternion==0.9.5
pyrsistent==0.15.4
python-dateutil==2.8.0
pythreejs==2.2.0
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.1.2
pyzmq==18.1.0
requests==2.23.0
scikit-image
scikit-learn==0.22.1
scipy==1.1.0
seaborn==0.9.0
Send2Trash==1.5.0
six==1.12.0
tensorboard==1.9.0
tensorflow-gpu==1.9.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
threevis==0.1.0.post25
torch==1.4.0
torchvision==0.5.0
tornado==6.0.3
tqdm==4.36.1
traitlets==4.3.2
traits==6.0.0
traitsui==6.1.3
traittypes==0.2.1
transforms==0.1
urllib3==1.25.8
vtk==8.1.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.6
widgetsnbextension==3.5.1
